backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.0",
    "pydantic>=2.11.7",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.2",
    "python-pptx>=0.6.23",
    "pymupdf>=1.23.0",
    "requests>=2.32.0",
]

[dependency-groups]
dev = [
    "reportlab>=4.4.2",
]
